Improved the SDIO interface definition for the WiBear11n module.
While the previous PAD_CTRL works, using the same PAD CTRL keeps keeps me sane.
The regulator configuration still has to be worked out, if we want to, but I rechecked the implementation in K&Ks Kernel source code, which only uses VGEN4 of the PMIC.
Since we support sdio interrupts, the device tree mirrors this now as well.