Added support for the ssd1305 chip as a framebuffer device.
Various configuration properties of the Solomon SSD1305/6/7 series are now available from the device tree to configure the Bolymin 12832BWRNH$ (which is based on the SSD1305) correctly. The initialization sequence for the display of the U100 is hardcoded.
You need to add fbcon=map:1 to your bootargs. Otherwise, the ssd1305 will host the framebuffer console, which makes the system almost unusable.